*ccl: detangle importccl and backupccl packages
Previously, the importccl package would rely on the backupccl package.
It didn't have a good reason for doing so. This lead to a dependency
cycle when any BACKUP test tried to run an IMPORT, leading to a lack of
testing coverage of the interop of the 2 behaviors. An alternative would
be to have tests that need to run both commands live inside importccl --
but that seemed like a big hack and further reinforced that there wasn't
a great reason for this dependency in the first place.

The reason for the dependency was around helper methods around
re-writing descriptors which were needed by both operations. This commit
introduces a new package, bulkccl, that houses the common helpers
between the 2 operations.

RowCount was also shared between the 2 packages. It was moved to
roachpb, where it now lives beside BulkOpSummary.

Release note: None
