+++ pkg/ccl/backupccl/BUILD.bazel
@@ -23,7 +23,6 @@ go_library(
         "split_and_scatter_processor.go",
         "system_schema.go",
         "targets.go",
-        "testutils.go",
     ],
     embed = [":backupccl_go_proto"],
     importpath = "github.com/cockroachdb/cockroach/pkg/ccl/backupccl",
@@ -32,6 +31,7 @@ go_library(
         "//pkg/base",
         "//pkg/build",
         "//pkg/ccl/backupccl/backupresolver",
+        "//pkg/ccl/bulkccl",
         "//pkg/ccl/storageccl",
         "//pkg/ccl/utilccl",
         "//pkg/featureflag",
@@ -61,7 +61,6 @@ go_library(
         "//pkg/sql/catalog/descs",
         "//pkg/sql/catalog/multiregion",
         "//pkg/sql/catalog/schemadesc",
-        "//pkg/sql/catalog/schemaexpr",
         "//pkg/sql/catalog/systemschema",
         "//pkg/sql/catalog/tabledesc",
         "//pkg/sql/catalog/typedesc",
@@ -87,7 +86,6 @@ go_library(
         "//pkg/storage",
         "//pkg/storage/cloud",
         "//pkg/storage/cloudimpl",
-        "//pkg/testutils",
         "//pkg/util/ctxgroup",
         "//pkg/util/encoding",
         "//pkg/util/hlc",
@@ -106,8 +104,6 @@ go_library(
         "@com_github_gogo_protobuf//jsonpb",
         "@com_github_gogo_protobuf//types",
         "@com_github_gorhill_cronexpr//:cronexpr",
-        "@com_github_lib_pq//oid",
-        "@com_github_stretchr_testify//require",
     ],
 )
 
@@ -136,6 +132,8 @@ go_test(
     deps = [
         "//pkg/base",
         "//pkg/blobs",
+        "//pkg/ccl/bulkccl",
+        "//pkg/ccl/importccl",
         "//pkg/ccl/kvccl",
         "//pkg/ccl/partitionccl",
         "//pkg/ccl/storageccl",
