+++ pkg/ccl/backupccl/restore_planning.go
@@ -10,12 +10,12 @@ package backupccl
 
 import (
 	"context"
-	"go/constant"
 	"net/url"
 	"path"
 	"sort"
 	"strconv"
 
+	"github.com/cockroachdb/cockroach/pkg/ccl/bulkccl"
 	"github.com/cockroachdb/cockroach/pkg/ccl/storageccl"
 	"github.com/cockroachdb/cockroach/pkg/ccl/utilccl"
 	"github.com/cockroachdb/cockroach/pkg/featureflag"
@@ -34,30 +34,23 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/dbdesc"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/schemadesc"
-	"github.com/cockroachdb/cockroach/pkg/sql/catalog/schemaexpr"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/systemschema"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/tabledesc"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/typedesc"
 	"github.com/cockroachdb/cockroach/pkg/sql/covering"
-	"github.com/cockroachdb/cockroach/pkg/sql/parser"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
 	"github.com/cockroachdb/cockroach/pkg/sql/privilege"
 	"github.com/cockroachdb/cockroach/pkg/sql/roleoption"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlerrors"
-	"github.com/cockroachdb/cockroach/pkg/sql/types"
 	"github.com/cockroachdb/cockroach/pkg/storage/cloud"
 	"github.com/cockroachdb/cockroach/pkg/storage/cloudimpl"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/tracing"
 	"github.com/cockroachdb/errors"
-	"github.com/lib/pq/oid"
 )
-
-// DescRewriteMap maps old descriptor IDs to new descriptor and parent IDs.
-type DescRewriteMap map[descpb.ID]*jobspb.RestoreDetails_DescriptorRewrite
 
 const (
 	restoreOptIntoDB                    = "into_db"
@@ -77,122 +70,6 @@ var featureRestoreEnabled = settings.RegisterBoolSetting(
 	"set to true to enable restore, false to disable; default is true",
 	featureflag.FeatureFlagEnabledDefault,
 ).WithPublic()
-
-// rewriteViewQueryDBNames rewrites the passed table's ViewQuery replacing all
-// non-empty db qualifiers with `newDB`.
-//
-// TODO: this AST traversal misses tables named in strings (#24556).
-func rewriteViewQueryDBNames(table *tabledesc.Mutable, newDB string) error {
-	stmt, err := parser.ParseOne(table.ViewQuery)
-	if err != nil {
-		return pgerror.Wrapf(err, pgcode.Syntax,
-			"failed to parse underlying query from view %q", table.Name)
-	}
-	// Re-format to change all DB names to `newDB`.
-	f := tree.NewFmtCtx(tree.FmtParsable)
-	f.SetReformatTableNames(func(ctx *tree.FmtCtx, tn *tree.TableName) {
-		// empty catalog e.g. ``"".information_schema.tables` should stay empty.
-		if tn.CatalogName != "" {
-			tn.CatalogName = tree.Name(newDB)
-		}
-		ctx.WithReformatTableNames(nil, func() {
-			ctx.FormatNode(tn)
-		})
-	})
-	f.FormatNode(stmt.AST)
-	table.ViewQuery = f.CloseAndGetString()
-	return nil
-}
-
-// rewriteTypesInExpr rewrites all explicit ID type references in the input
-// expression string according to rewrites.
-func rewriteTypesInExpr(expr string, rewrites DescRewriteMap) (string, error) {
-	parsed, err := parser.ParseExpr(expr)
-	if err != nil {
-		return "", err
-	}
-	ctx := tree.NewFmtCtx(tree.FmtSerializable)
-	ctx.SetIndexedTypeFormat(func(ctx *tree.FmtCtx, ref *tree.OIDTypeReference) {
-		newRef := ref
-		if rw, ok := rewrites[typedesc.UserDefinedTypeOIDToID(ref.OID)]; ok {
-			newRef = &tree.OIDTypeReference{OID: typedesc.TypeIDToOID(rw.ID)}
-		}
-		ctx.WriteString(newRef.SQLString())
-	})
-	ctx.FormatNode(parsed)
-	return ctx.CloseAndGetString(), nil
-}
-
-// rewriteSequencesInExpr rewrites all sequence IDs in the input expression
-// string according to rewrites.
-func rewriteSequencesInExpr(expr string, rewrites DescRewriteMap) (string, error) {
-	parsed, err := parser.ParseExpr(expr)
-	if err != nil {
-		return "", err
-	}
-	rewriteFunc := func(expr tree.Expr) (recurse bool, newExpr tree.Expr, err error) {
-		id, ok := schemaexpr.GetSeqIDFromExpr(expr)
-		if !ok {
-			return true, expr, nil
-		}
-		annotateTypeExpr, ok := expr.(*tree.AnnotateTypeExpr)
-		if !ok {
-			return true, expr, nil
-		}
-
-		rewrite, ok := rewrites[descpb.ID(id)]
-		if !ok {
-			return true, expr, nil
-		}
-		annotateTypeExpr.Expr = tree.NewNumVal(
-			constant.MakeInt64(int64(rewrite.ID)),
-			strconv.Itoa(int(rewrite.ID)),
-			false, /* negative */
-		)
-		return false, annotateTypeExpr, nil
-	}
-
-	newExpr, err := tree.SimpleVisit(parsed, rewriteFunc)
-	if err != nil {
-		return "", err
-	}
-	return newExpr.String(), nil
-}
-
-// rewriteSequencesInView walks the given viewQuery and
-// rewrites all sequence IDs in it according to rewrites.
-func rewriteSequencesInView(viewQuery string, rewrites DescRewriteMap) (string, error) {
-	rewriteFunc := func(expr tree.Expr) (recurse bool, newExpr tree.Expr, err error) {
-		id, ok := schemaexpr.GetSeqIDFromExpr(expr)
-		if !ok {
-			return true, expr, nil
-		}
-		annotateTypeExpr, ok := expr.(*tree.AnnotateTypeExpr)
-		if !ok {
-			return true, expr, nil
-		}
-		rewrite, ok := rewrites[descpb.ID(id)]
-		if !ok {
-			return true, expr, nil
-		}
-		annotateTypeExpr.Expr = tree.NewNumVal(
-			constant.MakeInt64(int64(rewrite.ID)),
-			strconv.Itoa(int(rewrite.ID)),
-			false, /* negative */
-		)
-		return false, annotateTypeExpr, nil
-	}
-
-	stmt, err := parser.ParseOne(viewQuery)
-	if err != nil {
-		return "", err
-	}
-	newStmt, err := tree.SimpleStmtVisit(stmt.AST, rewriteFunc)
-	if err != nil {
-		return "", err
-	}
-	return newStmt.String(), nil
-}
 
 // maybeFilterMissingViews filters the set of tables to restore to exclude views
 // whose dependencies are either missing or are themselves unrestorable due to
@@ -296,8 +173,8 @@ func allocateDescriptorRewrites(
 	descriptorCoverage tree.DescriptorCoverage,
 	opts tree.RestoreOptions,
 	intoDB string,
-) (DescRewriteMap, error) {
-	descriptorRewrites := make(DescRewriteMap)
+) (bulkccl.DescRewriteMap, error) {
+	descriptorRewrites := make(bulkccl.DescRewriteMap)
 	var overrideDB string
 	var renaming bool
 	if opts.IntoDB != nil {
@@ -988,352 +865,6 @@ func maybeUpgradeTableDescsInBackupManifests(
 	return nil
 }
 
-// rewriteDatabaseDescs rewrites all ID's in the input slice of
-// DatabaseDescriptors using the input ID rewrite mapping.
-func rewriteDatabaseDescs(databases []*dbdesc.Mutable, descriptorRewrites DescRewriteMap) error {
-	for _, db := range databases {
-		rewrite, ok := descriptorRewrites[db.ID]
-		if !ok {
-			return errors.Errorf("missing rewrite for database %d", db.ID)
-		}
-		db.ID = rewrite.ID
-
-		db.Version = 1
-		db.ModificationTime = hlc.Timestamp{}
-
-		// Rewrite the name-to-ID mapping for the database's child schemas.
-		newSchemas := make(map[string]descpb.DatabaseDescriptor_SchemaInfo)
-		for schemaName, schemaInfo := range db.Schemas {
-			if schemaInfo.Dropped {
-				continue
-			}
-			rewrite, ok := descriptorRewrites[schemaInfo.ID]
-			if !ok {
-				return errors.Errorf("missing rewrite for schema %d", db.ID)
-			}
-			newSchemas[schemaName] = descpb.DatabaseDescriptor_SchemaInfo{ID: rewrite.ID}
-		}
-		db.Schemas = newSchemas
-	}
-	return nil
-}
-
-// rewriteIDsInTypesT rewrites all ID's in the input types.T using the input
-// ID rewrite mapping.
-func rewriteIDsInTypesT(typ *types.T, descriptorRewrites DescRewriteMap) {
-	if !typ.UserDefined() {
-		return
-	}
-	// Collect potential new OID values.
-	var newOID, newArrayOID oid.Oid
-	if rw, ok := descriptorRewrites[typedesc.GetTypeDescID(typ)]; ok {
-		newOID = typedesc.TypeIDToOID(rw.ID)
-	}
-	if typ.Family() != types.ArrayFamily {
-		if rw, ok := descriptorRewrites[typedesc.GetArrayTypeDescID(typ)]; ok {
-			newArrayOID = typedesc.TypeIDToOID(rw.ID)
-		}
-	}
-	types.RemapUserDefinedTypeOIDs(typ, newOID, newArrayOID)
-	// If the type is an array, then we need to rewrite the element type as well.
-	if typ.Family() == types.ArrayFamily {
-		rewriteIDsInTypesT(typ.ArrayContents(), descriptorRewrites)
-	}
-}
-
-// rewriteTypeDescs rewrites all ID's in the input slice of TypeDescriptors
-// using the input ID rewrite mapping.
-func rewriteTypeDescs(types []*typedesc.Mutable, descriptorRewrites DescRewriteMap) error {
-	for _, typ := range types {
-		rewrite, ok := descriptorRewrites[typ.ID]
-		if !ok {
-			return errors.Errorf("missing rewrite for type %d", typ.ID)
-		}
-		// Reset the version and modification time on this new descriptor.
-		typ.Version = 1
-		typ.ModificationTime = hlc.Timestamp{}
-
-		typ.ID = rewrite.ID
-		typ.ParentSchemaID = maybeRewriteSchemaID(typ.ParentSchemaID, descriptorRewrites,
-			false /* isTemporaryDesc */)
-		typ.ParentID = rewrite.ParentID
-		for i := range typ.ReferencingDescriptorIDs {
-			id := typ.ReferencingDescriptorIDs[i]
-			if rw, ok := descriptorRewrites[id]; ok {
-				typ.ReferencingDescriptorIDs[i] = rw.ID
-			}
-		}
-		switch t := typ.Kind; t {
-		case descpb.TypeDescriptor_ENUM, descpb.TypeDescriptor_MULTIREGION_ENUM:
-			if rw, ok := descriptorRewrites[typ.ArrayTypeID]; ok {
-				typ.ArrayTypeID = rw.ID
-			}
-		case descpb.TypeDescriptor_ALIAS:
-			// We need to rewrite any ID's present in the aliased types.T.
-			rewriteIDsInTypesT(typ.Alias, descriptorRewrites)
-		default:
-			return errors.AssertionFailedf("unknown type kind %s", t.String())
-		}
-	}
-	return nil
-}
-
-// rewriteSchemaDescs rewrites all ID's in the input slice of SchemaDescriptors
-// using the input ID rewrite mapping.
-func rewriteSchemaDescs(schemas []*schemadesc.Mutable, descriptorRewrites DescRewriteMap) error {
-	for _, sc := range schemas {
-		rewrite, ok := descriptorRewrites[sc.ID]
-		if !ok {
-			return errors.Errorf("missing rewrite for schema %d", sc.ID)
-		}
-		// Reset the version and modification time on this new descriptor.
-		sc.Version = 1
-		sc.ModificationTime = hlc.Timestamp{}
-
-		sc.ID = rewrite.ID
-		sc.ParentID = rewrite.ParentID
-	}
-	return nil
-}
-
-func maybeRewriteSchemaID(
-	curSchemaID descpb.ID, descriptorRewrites DescRewriteMap, isTemporaryDesc bool,
-) descpb.ID {
-	// If the current schema is the public schema, then don't attempt to
-	// do any rewriting.
-	if curSchemaID == keys.PublicSchemaID && !isTemporaryDesc {
-		return curSchemaID
-	}
-	rw, ok := descriptorRewrites[curSchemaID]
-	if !ok {
-		return curSchemaID
-	}
-	return rw.ID
-}
-
-// RewriteTableDescs mutates tables to match the ID and privilege specified
-// in descriptorRewrites, as well as adjusting cross-table references to use the
-// new IDs. overrideDB can be specified to set database names in views.
-func RewriteTableDescs(
-	tables []*tabledesc.Mutable, descriptorRewrites DescRewriteMap, overrideDB string,
-) error {
-	for _, table := range tables {
-		tableRewrite, ok := descriptorRewrites[table.ID]
-		if !ok {
-			return errors.Errorf("missing table rewrite for table %d", table.ID)
-		}
-		// Reset the version and modification time on this new descriptor.
-		table.Version = 1
-		table.ModificationTime = hlc.Timestamp{}
-
-		if table.IsView() && overrideDB != "" {
-			// restore checks that all dependencies are also being restored, but if
-			// the restore is overriding the destination database, qualifiers in the
-			// view query string may be wrong. Since the destination override is
-			// applied to everything being restored, anything the view query
-			// references will be in the override DB post-restore, so all database
-			// qualifiers in the view query should be replaced with overrideDB.
-			if err := rewriteViewQueryDBNames(table, overrideDB); err != nil {
-				return err
-			}
-		}
-
-		table.ID = tableRewrite.ID
-		table.UnexposedParentSchemaID = maybeRewriteSchemaID(table.GetParentSchemaID(),
-			descriptorRewrites, table.IsTemporary())
-		table.ParentID = tableRewrite.ParentID
-
-		// Remap type IDs and sequence IDs in all serialized expressions within the TableDescriptor.
-		// TODO (rohany): This needs tests once partial indexes are ready.
-		if err := tabledesc.ForEachExprStringInTableDesc(table, func(expr *string) error {
-			newExpr, err := rewriteTypesInExpr(*expr, descriptorRewrites)
-			if err != nil {
-				return err
-			}
-			*expr = newExpr
-
-			newExpr, err = rewriteSequencesInExpr(*expr, descriptorRewrites)
-			if err != nil {
-				return err
-			}
-			*expr = newExpr
-			return nil
-		}); err != nil {
-			return err
-		}
-
-		// Walk view query and remap sequence IDs.
-		if table.IsView() {
-			viewQuery, err := rewriteSequencesInView(table.ViewQuery, descriptorRewrites)
-			if err != nil {
-				return err
-			}
-			table.ViewQuery = viewQuery
-		}
-
-		if err := catalog.ForEachNonDropIndex(table, func(indexI catalog.Index) error {
-			index := indexI.IndexDesc()
-			// Verify that for any interleaved index being restored, the interleave
-			// parent is also being restored. Otherwise, the interleave entries in the
-			// restored IndexDescriptors won't have anything to point to.
-			// TODO(dan): It seems like this restriction could be lifted by restoring
-			// stub TableDescriptors for the missing interleave parents.
-			for j, a := range index.Interleave.Ancestors {
-				ancestorRewrite, ok := descriptorRewrites[a.TableID]
-				if !ok {
-					return errors.Errorf(
-						"cannot restore table %q without interleave parent %d", table.Name, a.TableID,
-					)
-				}
-				index.Interleave.Ancestors[j].TableID = ancestorRewrite.ID
-			}
-			for j, c := range index.InterleavedBy {
-				childRewrite, ok := descriptorRewrites[c.Table]
-				if !ok {
-					return errors.Errorf(
-						"cannot restore table %q without interleave child table %d", table.Name, c.Table,
-					)
-				}
-				index.InterleavedBy[j].Table = childRewrite.ID
-			}
-			return nil
-		}); err != nil {
-			return err
-		}
-
-		// TODO(lucy): deal with outbound foreign key mutations here as well.
-		origFKs := table.OutboundFKs
-		table.OutboundFKs = nil
-		for i := range origFKs {
-			fk := &origFKs[i]
-			to := fk.ReferencedTableID
-			if indexRewrite, ok := descriptorRewrites[to]; ok {
-				fk.ReferencedTableID = indexRewrite.ID
-				fk.OriginTableID = tableRewrite.ID
-			} else {
-				// If indexRewrite doesn't exist, the user has specified
-				// restoreOptSkipMissingFKs. Error checking in the case the user hasn't has
-				// already been done in allocateDescriptorRewrites.
-				continue
-			}
-
-			// TODO(dt): if there is an existing (i.e. non-restoring) table with
-			// a db and name matching the one the FK pointed to at backup, should
-			// we update the FK to point to it?
-			table.OutboundFKs = append(table.OutboundFKs, *fk)
-		}
-
-		origInboundFks := table.InboundFKs
-		table.InboundFKs = nil
-		for i := range origInboundFks {
-			ref := &origInboundFks[i]
-			if refRewrite, ok := descriptorRewrites[ref.OriginTableID]; ok {
-				ref.ReferencedTableID = tableRewrite.ID
-				ref.OriginTableID = refRewrite.ID
-				table.InboundFKs = append(table.InboundFKs, *ref)
-			}
-		}
-
-		for i, dest := range table.DependsOn {
-			if depRewrite, ok := descriptorRewrites[dest]; ok {
-				table.DependsOn[i] = depRewrite.ID
-			} else {
-				// Views with missing dependencies should have been filtered out
-				// or have caused an error in maybeFilterMissingViews().
-				return errors.AssertionFailedf(
-					"cannot restore %q because referenced table %d was not found",
-					table.Name, dest)
-			}
-		}
-		for i, dest := range table.DependsOnTypes {
-			if depRewrite, ok := descriptorRewrites[dest]; ok {
-				table.DependsOnTypes[i] = depRewrite.ID
-			} else {
-				// Views with missing dependencies should have been filtered out
-				// or have caused an error in maybeFilterMissingViews().
-				return errors.AssertionFailedf(
-					"cannot restore %q because referenced type %d was not found",
-					table.Name, dest)
-			}
-		}
-		origRefs := table.DependedOnBy
-		table.DependedOnBy = nil
-		for _, ref := range origRefs {
-			if refRewrite, ok := descriptorRewrites[ref.ID]; ok {
-				ref.ID = refRewrite.ID
-				table.DependedOnBy = append(table.DependedOnBy, ref)
-			}
-		}
-
-		if table.IsSequence() && table.SequenceOpts.HasOwner() {
-			if ownerRewrite, ok := descriptorRewrites[table.SequenceOpts.SequenceOwner.OwnerTableID]; ok {
-				table.SequenceOpts.SequenceOwner.OwnerTableID = ownerRewrite.ID
-			} else {
-				// The sequence's owner table is not being restored, thus we simply
-				// remove the ownership dependency. To get here, the user must have
-				// specified 'skip_missing_sequence_owners', otherwise we would have
-				// errored out in allocateDescriptorRewrites.
-				table.SequenceOpts.SequenceOwner = descpb.TableDescriptor_SequenceOpts_SequenceOwner{}
-			}
-		}
-
-		// rewriteCol is a closure that performs the ID rewrite logic on a column.
-		rewriteCol := func(col *descpb.ColumnDescriptor) error {
-			// Rewrite the types.T's IDs present in the column.
-			rewriteIDsInTypesT(col.Type, descriptorRewrites)
-			var newUsedSeqRefs []descpb.ID
-			for _, seqID := range col.UsesSequenceIds {
-				if rewrite, ok := descriptorRewrites[seqID]; ok {
-					newUsedSeqRefs = append(newUsedSeqRefs, rewrite.ID)
-				} else {
-					// The referenced sequence isn't being restored.
-					// Strip the DEFAULT expression and sequence references.
-					// To get here, the user must have specified 'skip_missing_sequences' --
-					// otherwise, would have errored out in allocateDescriptorRewrites.
-					newUsedSeqRefs = []descpb.ID{}
-					col.DefaultExpr = nil
-					break
-				}
-			}
-			col.UsesSequenceIds = newUsedSeqRefs
-
-			var newOwnedSeqRefs []descpb.ID
-			for _, seqID := range col.OwnsSequenceIds {
-				// We only add the sequence ownership dependency if the owned sequence
-				// is being restored.
-				// If the owned sequence is not being restored, the user must have
-				// specified 'skip_missing_sequence_owners' to get here, otherwise
-				// we would have errored out in allocateDescriptorRewrites.
-				if rewrite, ok := descriptorRewrites[seqID]; ok {
-					newOwnedSeqRefs = append(newOwnedSeqRefs, rewrite.ID)
-				}
-			}
-			col.OwnsSequenceIds = newOwnedSeqRefs
-
-			return nil
-		}
-
-		// Rewrite sequence and type references in column descriptors.
-		for idx := range table.Columns {
-			if err := rewriteCol(&table.Columns[idx]); err != nil {
-				return err
-			}
-		}
-		for idx := range table.Mutations {
-			if col := table.Mutations[idx].GetColumn(); col != nil {
-				if err := rewriteCol(col); err != nil {
-					return err
-				}
-			}
-		}
-
-		// since this is a "new" table in eyes of new cluster, any leftover change
-		// lease is obviously bogus (plus the nodeID is relative to backup cluster).
-		table.Lease = nil
-	}
-	return nil
-}
-
 func errOnMissingRange(span covering.Range, start, end hlc.Timestamp) error {
 	return errors.Errorf(
 		"no backup covers time [%s,%s) for range [%s,%s) (or backups out of order)",
@@ -1825,16 +1356,16 @@ func doRestorePlan(
 
 	// We attempt to rewrite ID's in the collected type and table descriptors
 	// to catch errors during this process here, rather than in the job itself.
-	if err := RewriteTableDescs(tables, descriptorRewrites, intoDB); err != nil {
+	if err := bulkccl.RewriteTableDescs(tables, descriptorRewrites, intoDB); err != nil {
 		return err
 	}
-	if err := rewriteDatabaseDescs(databases, descriptorRewrites); err != nil {
+	if err := bulkccl.RewriteDatabaseDescs(databases, descriptorRewrites); err != nil {
 		return err
 	}
-	if err := rewriteSchemaDescs(schemas, descriptorRewrites); err != nil {
+	if err := bulkccl.RewriteSchemaDescs(schemas, descriptorRewrites); err != nil {
 		return err
 	}
-	if err := rewriteTypeDescs(types, descriptorRewrites); err != nil {
+	if err := bulkccl.RewriteTypeDescs(types, descriptorRewrites); err != nil {
 		return err
 	}
