+++ pkg/ccl/importccl/import_processor_test.go
@@ -21,7 +21,6 @@ import (
 
 	"github.com/cockroachdb/cockroach/pkg/base"
 	"github.com/cockroachdb/cockroach/pkg/blobs"
-	"github.com/cockroachdb/cockroach/pkg/ccl/backupccl"
 	"github.com/cockroachdb/cockroach/pkg/jobs"
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
 	"github.com/cockroachdb/cockroach/pkg/kv"
@@ -668,7 +667,7 @@ func TestCSVImportCanBeResumed(t *testing.T) {
 	jobCtx, cancelImport := context.WithCancel(ctx)
 	jobIDCh := make(chan jobspb.JobID)
 	var jobID jobspb.JobID = -1
-	var importSummary backupccl.RowCount
+	var importSummary roachpb.RowCount
 
 	registry.TestingResumerCreationKnobs = map[jobspb.Type]func(raw jobs.Resumer) jobs.Resumer{
 		// Arrange for our special job resumer to be
@@ -678,7 +677,7 @@ func TestCSVImportCanBeResumed(t *testing.T) {
 			resumer := raw.(*importResumer)
 			resumer.testingKnobs.ignoreProtectedTimestamps = true
 			resumer.testingKnobs.alwaysFlushJobProgress = true
-			resumer.testingKnobs.afterImport = func(summary backupccl.RowCount) error {
+			resumer.testingKnobs.afterImport = func(summary roachpb.RowCount) error {
 				importSummary = summary
 				return nil
 			}
@@ -776,7 +775,7 @@ func TestCSVImportMarksFilesFullyProcessed(t *testing.T) {
 	controllerBarrier, importBarrier := newSyncBarrier()
 
 	var jobID jobspb.JobID = -1
-	var importSummary backupccl.RowCount
+	var importSummary roachpb.RowCount
 
 	registry.TestingResumerCreationKnobs = map[jobspb.Type]func(raw jobs.Resumer) jobs.Resumer{
 		// Arrange for our special job resumer to be
@@ -785,7 +784,7 @@ func TestCSVImportMarksFilesFullyProcessed(t *testing.T) {
 			resumer := raw.(*importResumer)
 			resumer.testingKnobs.alwaysFlushJobProgress = true
 			resumer.testingKnobs.ignoreProtectedTimestamps = true
-			resumer.testingKnobs.afterImport = func(summary backupccl.RowCount) error {
+			resumer.testingKnobs.afterImport = func(summary roachpb.RowCount) error {
 				importSummary = summary
 				return nil
 			}
